+++ tcpclv4/fuzz/src/lib.rs
@@ -22,8 +22,8 @@ impl cla::Sink for MockSink {
         _peer_node: Option<&NodeId>,
         _peer_addr: Option<&cla::ClaAddress>,
         _stream: &mut dyn hardy_bpa::stream::Receiver<hardy_bpa::cla::Segment>,
-    ) -> cla::Result<()> {
-        Ok(())
+    ) -> cla::Result<cla::Acceptance> {
+        Ok(cla::Acceptance::Accepted)
     }
 
     async fn add_peer(